requirements.txt
abcvoting==2.10.0
absl-py==2.1.0
astunparse==1.6.3
async-generator==1.10
cachetools==5.5.2
certifi==2025.1.31
cffi==1.15.1
charset-normalizer==3.4.1
contourpy==1.2.1
cycler==0.12.1
Cython==0.29.37
Deprecated==1.2.14
distlib==0.3.8
filelock==3.14.0
flatbuffers==25.2.10
fonttools==4.53.0
fsspec==2024.6.0
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.71.0
gurobipy==11.0.2
h5py==3.13.0
idna==3.10
immutabledict==4.2.0
Jinja2==3.1.4
keras==2.13.1
kiwisolver==1.4.5
libclang==18.1.1
llvmlite==0.41.1
lz4==3.1.3
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.0
mip==1.15.0
mpmath==1.3.0
nashpy==0.0.40
networkx==3.3
numba==0.58.1
numpy==1.24.3
oauthlib==3.2.2
opt_einsum==3.4.0
ortools==9.10.4067
packaging==24.0
pillow==10.3.0
platformdirs==4.2.2
pref-voting==1.13.18
preflibtools==2.0.22
prefsampling==0.1.20
protobuf>=5.26.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.10.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
random2==1.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
seaborn==0.13.2
PyYAML==6.0.1
six==1.16.0
sympy==1.12.1
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow-estimator==2.13.0
termcolor==2.5.0
torch==2.3.1
torchvision==0.18.1
tqdm==4.67.1
typing_extensions>=4.8.0
tzdata==2024.1
urllib3==2.3.0
virtualenv==20.26.2
Werkzeug==3.1.3
wrapt==1.16.0
